fixed service report submit

+++ ticketing/app/Http/Controllers/TechnicianDashboardController.php
@@ -128,22 +128,21 @@ class TechnicianDashboardController extends Controller
 
     public function avatar(Request $request, $id)
     {
-        // Check if a file is uploaded
         if ($request->hasFile('avatar')) {
             $previousPath = $request->user()->getRawOriginal('avatar');
-            // Save the file to storage and get the path
             $link = Storage::put('/profile', $request->file('avatar'));
 
             $request->user()->update(['avatar' => $link]);
 
-            Storage::delete($previousPath);
+            if ($previousPath) {
+                Storage::delete($previousPath);
+            }
             
             return redirect()->back()
                 ->with('success', 'Profile Updated')
                 ->with('message', 'File uploaded successfully. Path: ' . $link);
         }
 
-        // If no file is uploaded, return an error message
         return redirect()->back()
             ->with('error', 'No File Uploaded');
     }
